#1q4y5x2[in progress] Add all new Entity related to Notification

+++ src/main/java/puretherapie/crm/api/v1/client/ClientInformation.java
@@ -35,7 +35,7 @@ public class ClientInformation {
     private boolean gender;
     private LocalDate birthday;
     private String phone;
-    private long idOrigin;
+    private int idOrigin;
 
     public void verify() throws ClientInformationVerificationException {
         Map<String, String> error = new HashMap<>();
