+++ src/main/java/puretherapie/crm/data/person/user/UserRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    User findByIdPerson(long idPerson);
+    User findByIdPerson(Integer idPerson);
 
     User findByUsername(String username);
